requirements.txt
mysql-connector-python==9.0.0
reportlab==4.2.2
PyPDF2==3.0.1
qrcode==7.4.2
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart
python-multipart
python-multipart
